requirements.txt
click==8.0.1
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-Cors==3.0.10
flask-mongoengine==1.0.0
flask_mongoengine==1.0.0
Flask-Scrypt==0.1.3.6
Flask-WTF==0.15.1
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
mongoengine==0.23.1
numpy==1.21.2
PyJWT==1.7.1
pymongo==3.12.0
python-dotenv==0.19.2
scrypt==0.8.18
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1
xlrd==2.0.1
WTForms==2.3.3
requests==2.28.1
